+++ modules/Content/Controller/Collection.php
@@ -73,7 +73,6 @@ class Collection extends App {
         }
 
         $fields = $model['fields'];
-
         $locales = $this->helper('locales')->locales();
 
         if (count($locales) == 1) {
@@ -83,6 +82,45 @@ class Collection extends App {
         }
 
         $this->helper('theme')->favicon(isset($model['icon']) && $model['icon'] ? $model['icon'] : 'content:assets/icons/collection.svg', $model['color'] ?? '#000');
+
+        return $this->render('content:views/collection/item.php', compact('model', 'fields', 'locales', 'item'));
+    }
+
+    public function clone($model = null, $id = null) {
+
+        if (!$model || !$id) {
+            return false;
+        }
+
+        $model = $this->module('content')->model($model);
+
+        if (!$model || $model['type'] != 'collection') {
+            return $this->stop(404);
+        }
+
+        if (!$this->isAllowed("content/{$model['name']}/create")) {
+            return $this->stop(401);
+        }
+
+        $item = $this->module('content')->item($model['name'], ['_id' => $id]);
+
+        if (!$item) {
+            return false;
+        }
+
+        // clean meta data
+        unset($item['_id'], $item['_created'], $item['_modified'], $item['_cby'], $item['_mby']);
+
+        $item['_state'] = 0;
+
+        $fields = $model['fields'];
+        $locales = $this->helper('locales')->locales();
+
+        if (count($locales) == 1) {
+            $locales = [];
+        } else {
+            $locales[0]['visible'] = true;
+        }
 
         return $this->render('content:views/collection/item.php', compact('model', 'fields', 'locales', 'item'));
     }